Hotfix: better peer connections handling + test reorganization
