+++ shell/tests/common/samples.rs
